# Refs #28459 -- Improved performance of loading DurationField on SQLite and MySQL.